refactor(client): break down SingleResourceEdit
Decompose the 575-LOC SingleResourceEdit (the largest function in the repo)
and its cyc-21 onSubmit handler into focused units, with no behavior change:

- Add a pure, testable buildResourcePayload(values, { isCostFromPlatform })
  helper that owns the per-field coalescing/branching, shrinking onSubmit
  below the cyc-10/cog-15 gate.
- Extract a useResourceEditForm hook (queries, options, change-tracked save
  form, provider-cost mirroring), kept in the route so the form state — and
  any unsaved edits — survives tab switches that unmount inactive panels.
- Move the details form JSX into a presentational DetailsTab sibling
  component plus a LevelSelectRow, leaving the route as orchestration only.
- Lean on the shared makeDuplicateHandler from useEditFormPage instead of the
  hand-rolled duplicate flow.

SingleResourceEdit drops from 575 to ~140 LOC and is no longer the top
large_functions entry; onSubmit clears the complexity gate.

Closes #456
