+++ packages/client/src/routes/resources.$id.edit.-components/-LevelSelectRow.tsx
@@ -1,0 +1,36 @@
+import { cn } from "@/lib/utils";
+import { changedFieldClass } from "@/utils";
+
+export type LevelValue = "" | "low" | "medium" | "high";
+
+export function LevelSelectRow({
+  label,
+  value,
+  onChange,
+  changed,
+}: {
+  label: string;
+  value: LevelValue;
+  onChange: (next: LevelValue) => void;
+  changed?: boolean;
+}) {
+  return (
+    <div className={cn("flex flex-col gap-1", changed && changedFieldClass)}>
+      <label className="text-xs font-medium text-muted-foreground">
+        {label}
+      </label>
+      <select
+        value={value}
+        onChange={e => onChange(e.target.value as LevelValue)}
+        className="
+          flex h-9 w-full rounded-md border bg-background px-3 py-1 text-sm
+        "
+      >
+        <option value="">—</option>
+        <option value="low">low</option>
+        <option value="medium">medium</option>
+        <option value="high">high</option>
+      </select>
+    </div>
+  );
+}
